Fixed includes and moved headers to proper directory.
