+++ Shared/Include/PartitionSegment.hpp
