+++ Shared/Include/PartitionSegmentBuilder.hpp
