+++ SharedLibs/Entities/Monsters/CMakeLists.txt
@@ -1,5 +1,5 @@
 cmake_minimum_required(VERSION 2.6.2)
 
-include_directories(../../../Shared/Libs/Interfaces ../../../Shared/Include)
+include_directories(../../../Shared/Libs/Interfaces ../../../Shared/Include ../../)
 
 add_subdirectory(DummyMonster)
